refactor(middleware): move Middleware from interfaces to types
