+++ source/types/Middleware.ts
@@ -1,6 +1,5 @@
 import { NextFunction } from 'express';
-import Request from './Request';
-import Response from './Response';
+import { Request, Response } from '../interfaces';
 
 type Middleware<T = any> = (
   req: Request,
