+++ source/types/index.ts
@@ -1,3 +1,4 @@
 export { default as MethodAttribute } from './MethodAttribute';
+export { default as Middleware } from '../types/Middleware';
 export { default as ProxyRequestHandler } from './ProxyRequestHandler';
 export { default as ResponseHeaders } from './ResponseHeaders';
